agent/requirements.txt
bedrock-agentcore 
strands-agents 
bedrock-agentcore-starter-toolkit
strands-agents-tools
python-dotenv
google-genai
